Remove need to run dev or beta version of Chrome.
This was only needed before because we used declarativeWebRequest
which has been in dev Chrome for years but never made it into stable.
The changes to work with the normal webRequest messes up the PizzaScript
API a little, and make things less effecient but at least we dont have
a weird dependency on non stable versions of Chrome.