Lame attempt at syncing the test result.
Really we need to either serialize this thing across the runner thread
and the websocket, or synchronise on every method. Also the websocket
library would be easier to use if it was just a single thread handling
the messages.  Right now it can still process the messages out of order
as far as i can tell.